Changed the JSON column in the trucks and orders tables to a one to many relationship with a properties table
This allows for SQL functionality in those properties, for example getting the different column names of a sheet.
You can still access these properties through a dictionary, so most implementations didn't have to change.

Also, changed the patch methods of the orders and trucks, so that setting a null value in the patch request deletes the property

NOTE: As always because of SQLite, `flask db downgrade` and `flask db upgrade` after pulling this commit